# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "rag_chatbot_backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/chat/sessions/{session_id}/messages/stream": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Create Message Stream",
        "operationId": "create_message_stream_api_v1_chat_sessions__session_id__messages_stream_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$",
              "description": "24-char hex Mongo ObjectId.",
              "examples": [
                "507f1f77bcf86cd799439011"
              ],
              "title": "Session Id"
            },
            "description": "24-char hex Mongo ObjectId."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Server-Sent Events (`text/event-stream`). Each event is `data: ` + one JSON object + two newlines. Events: `loading` (optional status), `delta` (`text` token), `done` (`message` = full assistant reply; optional `session_title` after the first reply when the session was auto-titled), or `error` (`detail`).",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                },
                "example": "data: {\"event\":\"loading\",\"message\":\"Thinking...\"}\n\ndata: {\"event\":\"delta\",\"text\":\"Hi\"}\n\ndata: {\"event\":\"done\",\"message\":\"Hi there\",\"session_title\":\"Greeting\"}\n\n"
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/sessions": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "List Sessions",
        "operationId": "list_sessions_api_v1_chat_sessions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChatSessionResponse"
                  },
                  "type": "array",
                  "title": "Response List Sessions Api V1 Chat Sessions Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Create Session",
        "operationId": "create_session_api_v1_chat_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/chat/sessions/{session_id}": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Session",
        "operationId": "get_session_api_v1_chat_sessions__session_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$",
              "description": "24-char hex Mongo ObjectId.",
              "examples": [
                "507f1f77bcf86cd799439011"
              ],
              "title": "Session Id"
            },
            "description": "24-char hex Mongo ObjectId."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSessionDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "chat"
        ],
        "summary": "Delete Session",
        "operationId": "delete_session_api_v1_chat_sessions__session_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$",
              "description": "24-char hex Mongo ObjectId.",
              "examples": [
                "507f1f77bcf86cd799439011"
              ],
              "title": "Session Id"
            },
            "description": "24-char hex Mongo ObjectId."
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/sessions/{session_id}/realtime/session": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Create Realtime Session",
        "description": "Mint an ephemeral OpenAI Realtime session for the browser.\n\nRetries transient OpenAI 5xx errors automatically. Enforces a per-user\nsliding-window rate limit driven by the audit collection.",
        "operationId": "create_realtime_session_api_v1_chat_sessions__session_id__realtime_session_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$",
              "description": "24-char hex Mongo ObjectId.",
              "examples": [
                "507f1f77bcf86cd799439011"
              ],
              "title": "Session Id"
            },
            "description": "24-char hex Mongo ObjectId."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RealtimeSessionMintResponse"
                }
              }
            }
          },
          "429": {
            "description": "Per-user mint rate limit exceeded."
          },
          "503": {
            "description": "OpenAI Realtime is unreachable after retries."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/sessions/{session_id}/realtime/tools/lookup_documentation": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Realtime Lookup Documentation",
        "operationId": "realtime_lookup_documentation_api_v1_chat_sessions__session_id__realtime_tools_lookup_documentation_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$",
              "description": "24-char hex Mongo ObjectId.",
              "examples": [
                "507f1f77bcf86cd799439011"
              ],
              "title": "Session Id"
            },
            "description": "24-char hex Mongo ObjectId."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LookupDocumentationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LookupDocumentationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/sessions/{session_id}/voice/commit": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Commit Voice Transcripts",
        "description": "Persist a completed OpenAI Realtime voice turn into the chat session.\n\nBehaviour:\n- Both transcripts are inserted via a single ``insert_many`` round trip; the\n  partial unique index on ``(session_id, client_turn_id, role)`` ensures retries\n  cannot create duplicates.\n- Idempotent: replays with the same ``client_turn_id`` return the originally\n  stored pair. Orphans from a previous crashed attempt are cleaned up first.\n- Title generation for the first assistant message is **fire-and-forget** so the\n  response returns immediately and the SPA can render the assistant bubble\n  without waiting on an LLM round trip.\n- Every commit is recorded in ``realtime_session_events`` for audit, abuse, and\n  cost-reconciliation queries.",
        "operationId": "commit_voice_transcripts_api_v1_chat_sessions__session_id__voice_commit_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$",
              "description": "24-char hex Mongo ObjectId.",
              "examples": [
                "507f1f77bcf86cd799439011"
              ],
              "title": "Session Id"
            },
            "description": "24-char hex Mongo ObjectId."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceCommitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Two persisted messages in order `[user, assistant]`, same shape as text chat (`ChatMessageResponse[]`) with `source = \"voice\"`. Title generation for the first assistant message runs as a background task and is **not** returned in this response; the SPA will see the title on the next `GET .../chat/sessions` call.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatMessageResponse"
                  },
                  "title": "Response Commit Voice Transcripts Api V1 Chat Sessions  Session Id  Voice Commit Post"
                }
              }
            }
          },
          "404": {
            "description": "Chat session not found or not owned by the caller."
          },
          "422": {
            "description": "Validation error (malformed session_id or transcripts exceeding length limits)."
          }
        }
      }
    },
    "/api/v1/chat/sessions/{session_id}/messages": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Create Message",
        "operationId": "create_message_api_v1_chat_sessions__session_id__messages_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{24}$",
              "description": "24-char hex Mongo ObjectId.",
              "examples": [
                "507f1f77bcf86cd799439011"
              ],
              "title": "Session Id"
            },
            "description": "24-char hex Mongo ObjectId."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatMessageResponse"
                  },
                  "title": "Response Create Message Api V1 Chat Sessions  Session Id  Messages Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatMessageCreate": {
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ChatMessageCreate"
      },
      "ChatMessageResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "role": {
            "$ref": "#/components/schemas/MessageRole"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "source": {
            "$ref": "#/components/schemas/MessageSource",
            "default": "text"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "user_id",
          "role",
          "content",
          "created_at"
        ],
        "title": "ChatMessageResponse"
      },
      "ChatSessionCreate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "type": "object",
        "title": "ChatSessionCreate"
      },
      "ChatSessionDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageResponse"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "title",
          "created_at",
          "updated_at"
        ],
        "title": "ChatSessionDetailResponse"
      },
      "ChatSessionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "title",
          "created_at",
          "updated_at"
        ],
        "title": "ChatSessionResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LookupDocumentationRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "LookupDocumentationRequest"
      },
      "LookupDocumentationResponse": {
        "properties": {
          "result": {
            "type": "string",
            "title": "Result"
          }
        },
        "type": "object",
        "required": [
          "result"
        ],
        "title": "LookupDocumentationResponse",
        "description": "Pass verbatim to OpenAI Realtime as the function tool output (plain text)."
      },
      "MessageRole": {
        "type": "string",
        "enum": [
          "user",
          "assistant",
          "system"
        ],
        "title": "MessageRole"
      },
      "MessageSource": {
        "type": "string",
        "enum": [
          "text",
          "voice"
        ],
        "title": "MessageSource",
        "description": "Where a stored message originated. Useful for analytics and debugging."
      },
      "RealtimeClientSecret": {
        "properties": {
          "value": {
            "type": "string",
            "title": "Value"
          },
          "expires_at": {
            "type": "integer",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "value",
          "expires_at"
        ],
        "title": "RealtimeClientSecret"
      },
      "RealtimeSessionMintResponse": {
        "properties": {
          "chat_session_id": {
            "type": "string",
            "title": "Chat Session Id"
          },
          "openai_session_id": {
            "type": "string",
            "title": "Openai Session Id"
          },
          "client_secret": {
            "$ref": "#/components/schemas/RealtimeClientSecret"
          },
          "model": {
            "type": "string",
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "chat_session_id",
          "openai_session_id",
          "client_secret",
          "model"
        ],
        "title": "RealtimeSessionMintResponse",
        "description": "Ephemeral Realtime credentials for the browser; chat_session_id ties persistence to Mongo."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VoiceCommitRequest": {
        "properties": {
          "user_transcript": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "User Transcript"
          },
          "assistant_transcript": {
            "type": "string",
            "maxLength": 16000,
            "minLength": 1,
            "title": "Assistant Transcript"
          },
          "client_turn_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Turn Id"
          },
          "openai_response_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Openai Response Id"
          }
        },
        "type": "object",
        "required": [
          "user_transcript",
          "assistant_transcript"
        ],
        "title": "VoiceCommitRequest",
        "description": "Body for ``POST .../voice/commit``.\n\n``client_turn_id`` makes retries idempotent: if the same id is replayed for the same\nchat session, the server returns the already-persisted user+assistant pair instead of\ninserting duplicates. ``openai_response_id`` is stored as a debug breadcrumb only."
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}